refactor all reusable constants and functions into the src/utils/ directory
