+++ src/app-bundles/bundle-utils.js
@@ -34,46 +34,3 @@ export function getRestUrl( liveUrl, mockUrl, mockOverrideFlag ) {
   else if( liveUrl.startsWith( "http" ) ) return liveUrl;
   else return `${ baseUrl }${ liveUrl }`;
 }
-
-export const arrayToObj = (array, uid) => {
-  const obj = {};
-  array.forEach((item) => {
-    obj[item[uid] || 0] = item;
-  });
-  return obj;
-};
-
-/**
- * Returns the number of milliseconds of the given time interval
- * @param {*} interval string representation of the time interval
- */
-export const getIntervalTime = (interval) => {
-  switch (interval) {
-    case "PT15M": // 15 minutes
-      return 900000;
-    case "PT30M":
-      return 1800000;
-    case "PT1H":  // 1 hour
-      return 3600000;
-    case "P1D":
-      return TIME.DAY;
-    case "P1M": // 1 month
-      return TIME.MONTH;
-    default:
-      return 0;
-  }
-};
-
-// number of milliseconds in a hour, day, week, month, and year
-export const TIME = {
-  DAY: 86400000, // (24 hours) * (60 minutes/hour) * (60 seconds/minute) * (1000 milliseconds/second)
-  WEEK: 604800000, // (7 days) * (24 hours/day) * (60 minutes/hour) * (60 seconds/minute) * (1000 milliseconds/second)
-  MONTH: 2592000000, // (30 days) * (24 hours/day) * (60 minutes/hour) * (60 seconds/minute) * (1000 milliseconds/second)
-  YEAR: 31536000000, // (365 days) * (24 hours/day) * (60 minutes/hour) * (60 seconds/minute) * (1000 milliseconds/second)
-};
-
-/**
- * Receives javascript Date object and converts it to string in the format yyyy-mm-dd
- * @param {Date} date the date to convert
- */
-export const dateToString = date => date && date.toISOString().substring(0, 10);
