+++ src/utils/constants.js
@@ -1,0 +1,7 @@
+// number of milliseconds in a hour, day, week, month, and year
+export const TIME = {
+  DAY: 86400000, // (24 hours) * (60 minutes/hour) * (60 seconds/minute) * (1000 milliseconds/second)
+  WEEK: 604800000, // (7 days) * (24 hours/day) * (60 minutes/hour) * (60 seconds/minute) * (1000 milliseconds/second)
+  MONTH: 2592000000, // (30 days) * (24 hours/day) * (60 minutes/hour) * (60 seconds/minute) * (1000 milliseconds/second)
+  YEAR: 31536000000, // (365 days) * (24 hours/day) * (60 minutes/hour) * (60 seconds/minute) * (1000 milliseconds/second)
+};
